Move logger module to acquisition.common
* includes deployment
