+++ src/acquisition/covidcast/file_archiver.py
@@ -6,7 +6,7 @@ import os
 import shutil
 
 # first party
-from delphi.epidata.acquisition.covidcast.logger import get_structured_logger
+from delphi.epidata.acquisition.common.logger import get_structured_logger
 
 class FileArchiver:
   """Archives files by moving and compressing."""
